refactor(client): move store out of registry folder
